+++ RELEASES.md
@@ -459,3 +459,13 @@ Based on:
 - [typescript v0.0.50] .
 ### Releases
 - [NPM v0.0.50] https://www.npmjs.com/package/cribl-control-plane/v/0.0.50 - .
+
+## 2025-09-08 17:51:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.606.0 (2.686.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.0.51-alpha.1] .
+### Releases
+- [NPM v0.0.51-alpha.1] https://www.npmjs.com/package/cribl-control-plane/v/0.0.51-alpha.1 - .